+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.0] - 2022-06-26
+### Added
+- Requires LiteNinja-Utils
+- HSL Color Space
+- HSV Color Space
+
+
+
 ## [0.0.1] - 2022-06-17
 Empty project